% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_longtests_rcolet.R
\name{internal_longtests_roclet_clean}
\alias{internal_longtests_roclet_clean}
\title{Clean up longtest files generated by biocroxytest}
\usage{
internal_longtests_roclet_clean(testfiles)
}
\arguments{
\item{testfiles}{A character vector representing the paths of the test files
to check and clean.}
}
\description{
This is an internal function that cleans up test files generated by
biocroxytest. It checks each file in the provided list of test files. If a
file was generated by biocroxytest, it will be removed. If a file was not
generated by biocroxytest, a message will be printed to inform the user, and
the file will not be removed.
}
\keyword{internal}
