% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_longtests_rcolet.R
\name{internal_longtests_roclet_output}
\alias{internal_longtests_roclet_output}
\title{Write long test results to files}
\usage{
internal_longtests_roclet_output(
  results,
  base_path,
  prefix = "test-biocroxytest"
)
}
\arguments{
\item{results}{A list of results.}

\item{base_path}{The base path where the files will be written.}

\item{prefix}{The prefix that will be used to name the files.}
}
\value{
A list of file paths.
}
\description{
This function takes three arguments: results, base_path, and prefix. The
function iterates over the elements of results and writes the content of each
element to a file in the directory specified by base_path and prefix. If the
file was not created by biocroxytest, the function informs the user that some
files were not removed.
}
\keyword{internal}
