% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/made_by.R
\name{made_by_biocroxytest}
\alias{made_by_biocroxytest}
\title{Check if a file is generated by biocroxytest}
\usage{
made_by_biocroxytest(path)
}
\arguments{
\item{path}{A character string representing the path of the file to check.}
}
\value{
TRUE if all checks pass, FALSE otherwise.
}
\description{
This is an internal function that checks if a file at a given path is
generated by biocroxytest. It checks if the file exists, if the name matches
the pattern "test-biocroxytest-.*\\.R$", and if the header of the file starts
with "Generated by biocroxytest". If the file exists and the name matches the
pattern but the header does not start with "Generated by biocroxytest", the
function will return FALSE.
}
\keyword{internal}
