% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_longtests.R
\name{use_longtests}
\alias{use_longtests}
\title{Sets up overall longtests infrastructure}
\usage{
use_longtests()
}
\value{
Invisible
}
\description{
Creates longtests/testthat directory and longtests/testthat.R file.
}
\examples{
# Create the longtests directory and .BBSoptions file
use_longtests()

# Remove the longtests directory and .BBSoptions file
unlink(file.path(".BBSoptions"), recursive = TRUE, force = TRUE)
unlink(file.path("longtests"), recursive = TRUE, force = TRUE)
}
