% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_longtests.R
\name{verify_longtests_used}
\alias{verify_longtests_used}
\title{Check if longtests are used in the package}
\usage{
verify_longtests_used(base_path = ".")
}
\arguments{
\item{base_path}{The base path of the package.}
}
\value{
TRUE if "longtests" directory exists, FALSE otherwise.
}
\description{
This function checks if the directory "longtests" exists in the current
package. If it does not exist, the function will throw an error message and
suggest setting it up using the \code{biocroxytest::use_longtests()} function.
}
