% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_longtests.R
\name{use_longtests}
\alias{use_longtests}
\title{Sets up overall longtests infrastructure}
\usage{
use_longtests()
}
\value{
Invisible
}
\description{
Creates longtests/testthat directory and longtests/testthat.R file.
}
